Updated smartmirror.py and added config.txt
smartmirror.py will not look to read from a config.txt file for modules to have shown.  If the value in the config file is 0 then that module will not be shown, if 1 it will be.  If config.txt is 0 1 1 it will not show the clock, but will show weather and news.  If it is 1 1 0 the clock and weather are shown but not news. etc.